Cleanup implementation

+++ x/stake/keeper_keys.go
@@ -32,10 +32,10 @@ func GetCandidateKey(addr sdk.Address) []byte {
 }
 
 // get the key for the validator used in the power-store
-func GetValidatorKey(addr sdk.Address, power sdk.Rat, height uint64, cdc *wire.Codec) []byte {
+func GetValidatorKey(addr sdk.Address, power sdk.Rat, height int64, cdc *wire.Codec) []byte {
 	powerBytes := []byte(power.ToLeftPadded(maxDigitsForAccount)) // power big-endian (more powerful validators first)
 	heightBytes := make([]byte, 8)
-	binary.LittleEndian.PutUint64(heightBytes, height) // height little-endian (older validators first)
+	binary.LittleEndian.PutUint64(heightBytes, uint64(height)) // height little-endian (older validators first)
 	return append(ValidatorsKey, append(powerBytes, append(heightBytes, addr.Bytes()...)...)...)
 }
